requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.2.7
django-appconf==1.1.0
django-celery-beat==2.8.1
django-extensions==4.1
django-imagekit==6.0.0
django-redis==6.0.0
django-taggit==6.1.0
django-timezone-field==7.1
djangorestframework==3.16.1
kombu==5.5.4
packaging==25.0
pilkit==3.0
pillow==12.0.0
prompt_toolkit==3.0.52
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
